Added heads extraction utility along with all input files
Synced all the heads (modelwide and lake) utility files with the NFSEG Reg Tool into the formerly archived directory in the same repository. The archived directory was renamed and moved to this repository.

This directory contains the complete set of files needed to extract heads from an NFSEG model heads output file. Both the original version of the utility from Wei that extracts UFA lake heads and the updated version that extracts modelwide and lake heads for all model layers are included.