Fix the year string in the CRH script
Fixed the year string used to describe and locate a component in the data. The year string was previously changed to include the full 4-digit year, but the first two digits were already written manually in a string describing the data source. Removed the first two digits in the string since it is now included in the year string